Avanves periodos y procesos electivos

+++ backend/src/services/periodo.service.js
@@ -1,13 +1,14 @@
 "use strict";
 const Periodo = require("../models/periodo.model.js");
-const { PERIODOS } = require("../utils/constants");
+const Proceso = require("../models/proceso.model.js");
+const Constants = require("../constants/periodos.constants.js");
 const { handleError } = require("../utils/errorHandler");
 
 async function getPeriodos() {
   try {
     const periodos = await Periodo.find();
-    if (!periodos)
-      return [null, "No hay periodos o etapas electivas registradas"];
+    if (periodos.length === 0)
+      return [null, "No hay periodos electivos registrados"];
 
     return [periodos, null];
   } catch (error) {
@@ -17,25 +18,49 @@ async function getPeriodos() {
 
 async function createPeriodo(periodo) {
   try {
-    const { nombre_etapa, fecha_inicio } = periodo;
+    const { nombre_etapa, fechaInicio, procesoId } = periodo;
 
-    //verificar si el periodo electivo ya existe
-    const periodoFound = await Periodo.findOne({ nombre_etapa });
-    if (periodoFound) return [null, "El periodo ya existe"];
+    let fechaInicioDate = new Date(fechaInicio);
 
-    const duracion = PERIODOS.find(
+    const duracion = Constants.find(
       (periodo) => periodo.nombre_etapa === nombre_etapa,
     ).duracion;
 
-    const numero_etapa = PERIODOS.find(
+    const numero_etapa = Constants.find(
       (periodo) => periodo.nombre_etapa === nombre_etapa,
     ).numero_etapa;
 
+    const proceso = await Proceso.findById(procesoId);
+    if (!proceso) return [null, "El proceso no existe"];
+
+    console.log("El periodo `${nombre_etapa}` ya existe dentro del proceso: '${proceso.nombre}'");
+
+    // Verificar si el periodo electivo ya existe dentro del proceso
+    const periodoFound = await Periodo.findOne({
+      nombre_etapa: nombre_etapa,
+      procesoId: procesoId,
+    });
+    if (periodoFound)
+      return [
+        null,
+        "El periodo `${nombre_etapa}` ya existe dentro del proceso: '${proceso.nombre}'",
+      ];
+
+
+    //verificar secuencia de etapas
+    if (numero_etapa !== 1) {
+      const periodoPrevio = await Periodo.findOne({
+        numero_etapa: numero_etapa - 1,
+      });
+
+      if (!periodoPrevio) return [null, "La secuencia de etapas es incorrecta"];
+    }
+
     const newPeriodo = new Periodo({
       nombre_etapa,
-      fecha_inicio,
-      fecha_fin: new Date(
-        fecha_inicio.getTime() + duracion * 24 * 60 * 60 * 1000,
+      fechaInicio: fechaInicioDate,
+      fechaFin: new Date(
+        fechaInicioDate.getTime() + duracion * 24 * 60 * 60 * 1000,
       ),
       duracion,
       numero_etapa,
@@ -52,17 +77,27 @@ async function createPeriodo(periodo) {
 
 async function updatePeriodo(id, periodo) {
   try {
-    const { nombre_etapa, fecha_inicio, duracion } = periodo;
+    const { nombre_etapa, fechaInicio } = periodo;
+    let fechaInicioDate = new Date(fechaInicio);
+
+    const duracion = Constants.find(
+      (periodo) => periodo.nombre_etapa === nombre_etapa,
+    ).duracion;
+
+    const numero_etapa = Constants.find(
+      (periodo) => periodo.nombre_etapa === nombre_etapa,
+    ).numero_etapa;
 
     const updatedPeriodo = await Periodo.findByIdAndUpdate(
       id,
       {
         nombre_etapa,
-        fecha_inicio,
-        fecha_fin: new Date(
-          fecha_inicio.getTime() + duracion * 24 * 60 * 60 * 1000,
+        fechaInicioDate,
+        fechaFin: new Date(
+          fechaInicioDate.getTime() + duracion * 24 * 60 * 60 * 1000,
         ),
         duracion,
+        numero_etapa,
       },
       { new: true },
     );
@@ -75,6 +110,9 @@ async function updatePeriodo(id, periodo) {
 
 async function deletePeriodo(id) {
   try {
+    const periodoFound = await Periodo.findById(id);
+    if (!periodoFound) return [null, "El periodo no existe"];
+
     const deletedPeriodo = await Periodo.findByIdAndDelete(id);
 
     return [deletedPeriodo, null];
